backend/src/requirements/requirements-dev.txt
-r requirements-prod.txt
django-cors-headers==4.1.0